feat: helper scripts to debug validation allow list (#461)
The upgrade went badly because it wasn't saving the SSO beacon to the
upgrade file, so a new beacon was deployed and used in the factory.
The new beacon wasn't in the validation allow list, so any transactions
using the proxy failed which means all accounts deployed via that
factory were bad.

Upgrading the factory helps, but we also need to upgrade the sso account
beacon to the latest implemention. This can't fix existing accounts,
but can work for new accounts.

The factory upgrade might not have worked, so we need to deploy another
paymaster to go with a new factory that deploys the allowed SSO beacon
address.

What a mess and I still can't get it working in PR...